production-requirements.txt
aiohttp==3.7.4.post0
async-timeout==3.0.1
atomicwrites==1.4.0
attrs==21.2.0
certifi==2021.5.30
chardet==4.0.0
colorama==0.4.4
discord.py==1.7.3
dpytest==0.5.3
emoji==1.4.2
idna==3.2
iniconfig==1.1.1
mock==4.0.3
more-itertools==8.8.0
multidict==5.1.0
packaging==21.0
parsedatetime==2.6
pluggy==0.13.1
py==1.10.0
pyparsing==2.4.7
pysqlcipher3==1.0.4
pytest==6.2.4
pytest-asyncio==0.15.1
pytest-ordering==0.6
python-dotenv==0.19.0
requests==2.26.0
six==1.16.0
sqlalchemy==1.4.25
toml==0.10.2
twitchAPI==2.4.1
urllib3==1.26.6
wcwidth==0.2.5
websockets==9.1
yarl==1.6.3
